Refactor VisibleNode as an interface instead of a class
