+++ alchemist/alchemist-smartcam/src/main/kotlin/it/unibo/alchemist/model/interfaces/VisibleNode.kt
@@ -1,40 +1,16 @@
-package it.unibo.alchemist.model.smartcam
-
-import it.unibo.alchemist.model.interfaces.Environment
-import it.unibo.alchemist.model.interfaces.Node
-import it.unibo.alchemist.model.interfaces.Position
+package it.unibo.alchemist.model.interfaces
 
 /**
  * Represents a node seen by the [it.unibo.alchemist.model.implementations.actions.See] action.
  */
-data class VisibleNode<T, P : Position<P>>(
+interface VisibleNode<T, P : Position<P>> {
     /**
      * The node seen.
      */
-    val node: Node<T>,
-    private val env: Environment<T, P>
-) {
+    val node: Node<T>
 
     /**
-     * Identifier.
+     * The position of the node
      */
-    val id = node.id
-
-    /**
-     * Absolute position.
-     */
-    val position: P = env.getPosition(node)
-
-    override fun equals(other: Any?): Boolean {
-        if (other is VisibleNode<*, *>) {
-            return other.id == id
-        }
-        return false
-    }
-
-    override fun hashCode() =
-        id.hashCode()
-
-    override fun toString() =
-        "Visible#$id"
+    val position: P
 }
